Move `BrowserData`
