+++ crates/swc_ecma_preset_env/src/lib.rs
@@ -7,6 +7,7 @@ use std::path::PathBuf;
 use anyhow::{Context, Error};
 use dashmap::DashMap;
 use once_cell::sync::Lazy;
+pub use preset_env_base::BrowserData;
 use serde::Deserialize;
 use st_map::StaticMap;
 use swc_atoms::{js_word, JsWord};
@@ -263,46 +264,6 @@ where
             excludes: excluded_modules,
         }
     )
-}
-
-/// A map without allocation.
-#[derive(Debug, Default, Deserialize, Clone, Copy, StaticMap)]
-#[serde(deny_unknown_fields)]
-pub struct BrowserData<T: Default> {
-    #[serde(default)]
-    pub chrome: T,
-    #[serde(default)]
-    pub and_chr: T,
-    #[serde(default)]
-    pub and_ff: T,
-    #[serde(default)]
-    pub op_mob: T,
-    #[serde(default)]
-    pub ie: T,
-    #[serde(default)]
-    pub edge: T,
-    #[serde(default)]
-    pub firefox: T,
-    #[serde(default)]
-    pub safari: T,
-    #[serde(default)]
-    pub node: T,
-    #[serde(default)]
-    pub ios: T,
-    #[serde(default)]
-    pub samsung: T,
-    #[serde(default)]
-    pub opera: T,
-    #[serde(default)]
-    pub android: T,
-    #[serde(default)]
-    pub electron: T,
-    #[serde(default)]
-    pub phantom: T,
-    #[serde(default)]
-    pub opera_mobile: T,
-    #[serde(default)]
-    pub rhino: T,
 }
 
 #[derive(Debug)]
